+++ hss_okta_flutter/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.2.2
+* Made OktaAuth private
+* Made OktaConfig Private
+* Added getter for OktaAuth
+* Added getter for OktaConfig
+* Added AuthStateManager
+* Added TokenManager
+* Added Helper methods for OktaAuth
+
 # 1.1.1
 * Updated readme.md